+++ inc/sortsmill/math/polyspline/eval.h
@@ -31,17 +31,17 @@ extern "C"
 #endif
 
 /* Polynomial evaluation in the monomial basis. */
-double eval_f64_mono (unsigned int degree, int stride,
+double eval_f64_mono (size_t degree, ssize_t stride,
                       const double *spline, double t);
-SCM scm_c_eval_mono (unsigned int degree, int stride, const SCM *spline, SCM t);
+SCM scm_c_eval_mono (size_t degree, ssize_t stride, const SCM *spline, SCM t);
 SCM scm_eval_f64_mono (SCM vector, SCM t);
 SCM scm_eval_scm_mono (SCM vector, SCM t);
 
 /* ‘Fast’ evaluation of polynomials in Bernstein form, by the
    reasonably stable algorithm of Schumaker and Volk. */
-double eval_f64_bern_schumaker_volk (unsigned int deg, int stride,
+double eval_f64_bern_schumaker_volk (size_t deg, ssize_t stride,
                                      const double *spline, double t);
-SCM scm_c_eval_bern_schumaker_volk (unsigned int deg, int stride,
+SCM scm_c_eval_bern_schumaker_volk (size_t deg, ssize_t stride,
                                     const SCM *spline, SCM t);
 SCM scm_eval_f64_bern_schumaker_volk (SCM splines, SCM t);
 SCM scm_eval_scm_bern_schumaker_volk (SCM splines, SCM t);
@@ -49,18 +49,18 @@ SCM scm_eval_scm_bern_schumaker_volk (SCM splines, SCM t);
 /* Evaluation of polynomials in Bernstein form, by the very stable
    algorithm of De Casteljau. See
    http://en.wikipedia.org/wiki/De_Casteljau%27s_algorithm */
-double eval_f64_bern_de_casteljau (unsigned int deg, int stride,
+double eval_f64_bern_de_casteljau (size_t deg, ssize_t stride,
                                    const double *spline, double t);
-SCM scm_c_eval_bern_de_casteljau (unsigned int deg, int stride,
+SCM scm_c_eval_bern_de_casteljau (size_t deg, ssize_t stride,
                                   const SCM *spline, SCM t);
 SCM scm_eval_f64_bern_de_casteljau (SCM splines, SCM t);
 SCM scm_eval_scm_bern_de_casteljau (SCM splines, SCM t);
 
 /* ‘Fast’ evaluation of polynomials in scaled Bernstein form, by the
    reasonably stable algorithm of Schumaker and Volk. */
-double eval_f64_sbern_schumaker_volk (unsigned int deg, int stride,
+double eval_f64_sbern_schumaker_volk (size_t deg, ssize_t stride,
                                       const double *spline, double t);
-SCM scm_c_eval_sbern_schumaker_volk (unsigned int deg, int stride,
+SCM scm_c_eval_sbern_schumaker_volk (size_t deg, ssize_t stride,
                                      const SCM *spline, SCM t);
 SCM scm_eval_f64_sbern_schumaker_volk (SCM vector, SCM t);
 SCM scm_eval_scm_sbern_schumaker_volk (SCM vector, SCM t);
@@ -68,18 +68,17 @@ SCM scm_eval_scm_sbern_schumaker_volk (SCM vector, SCM t);
 /* Evaluation of polynomials in scaled Bernstein form, by the very
    stable algorithm of De Casteljau. See
    http://en.wikipedia.org/wiki/De_Casteljau%27s_algorithm */
-double eval_f64_sbern_de_casteljau (unsigned int deg, int stride,
+double eval_f64_sbern_de_casteljau (size_t deg, ssize_t stride,
                                     const double *spline, double t);
-SCM scm_c_eval_sbern_de_casteljau (unsigned int deg, int stride,
+SCM scm_c_eval_sbern_de_casteljau (size_t deg, ssize_t stride,
                                    const SCM *spline, SCM t);
 SCM scm_eval_f64_sbern_de_casteljau (SCM splines, SCM t);
 SCM scm_eval_scm_sbern_de_casteljau (SCM splines, SCM t);
 
 /* Polynomial evaluation in the s-power basis of Sánchez-Reyes. */
-double eval_f64_spower (unsigned int degree, int stride,
+double eval_f64_spower (size_t degree, ssize_t stride,
                         const double *spline, double t);
-SCM scm_c_eval_spower (unsigned int degree, int stride,
-                       const SCM *spline, SCM t);
+SCM scm_c_eval_spower (size_t degree, ssize_t stride, const SCM *spline, SCM t);
 SCM scm_eval_f64_spower (SCM vector, SCM t);
 SCM scm_eval_scm_spower (SCM vector, SCM t);
